fixed requests and approvals

+++ src/main/java/MojaChata/pl/app/controller/RequestController.java
@@ -10,7 +10,9 @@ import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @Controller
@@ -26,10 +28,13 @@ public class RequestController {
 
     @GetMapping("/unrequest")
     public String unsendRequest(@SessionAttribute(value = "loggedInUser", required = false) User login,
-                                @RequestParam("cottageId") long cottageId, Model model) {
-        Request request = requestRepository.findByCustomerIdAndCottageId(login.getId(), cottageId);
-        RequestApproval requestApproval = requestApprovalsRepository.findByRequestId(request.getId());
-        requestApprovalsRepository.delete(requestApproval);
+                                @RequestParam("requestId") long requestId, Model model) {
+        Request request = requestRepository.findById(requestId)
+                .orElseThrow(() -> new IllegalArgumentException("Invalid request Id:" + requestId));;
+        RequestApproval requestApproval = requestApprovalsRepository.findByRequestId(requestId);
+        if (requestApproval != null) {
+            requestApprovalsRepository.delete(requestApproval);
+        }
         requestRepository.delete(request);
         return "redirect:/reservations";
     }
@@ -43,7 +48,7 @@ public class RequestController {
 
         Cottage cottage = cottageRepository.findById(cottageId)
                 .orElseThrow(() -> new RuntimeException("Invalid cottage Id: " + cottageId));
-        if (requestRepository.existsByCustomerIdAndCottageId(login.getId(), cottage.getId())) {
+        if (requestService.existsByCustomerIdAndCottageIdAndNoApproval(login.getId(), cottage.getId())) {
             return "redirect:/reservations";
         }
 
@@ -86,17 +91,29 @@ public class RequestController {
     @GetMapping("/reservations")
     public String showReservations(@SessionAttribute(value = "loggedInUser", required = false) User login,
                                    @RequestParam(value = "status", required = false) String status, Model model) {
-        List<RequestApproval> approvals;
+        List<Request> requests = requestRepository.findByCustomerId(login.getId());
+        Map<Long, Integer> requestStatus = new HashMap<>();
 
         if ("approved".equalsIgnoreCase(status)) {
-            approvals = requestApprovalsRepository.findByIsApprovedTrueAndRequest_CustomerId(login.getId());
+            requests = requests.stream()
+                    .filter(request -> requestService.existsByRequestIdAndIsApproved(request.getId(), true))
+                    .collect(Collectors.toList());
+        } else if ("rejected".equalsIgnoreCase(status)) {
+            requests = requests.stream()
+                    .filter(request -> requestService.existsByRequestIdAndIsApproved(request.getId(), false))
+                    .collect(Collectors.toList());
         } else if ("pending".equalsIgnoreCase(status)) {
-            approvals = requestApprovalsRepository.findByIsApprovedFalseAndRequest_CustomerId(login.getId());
-        } else {
-            approvals = requestApprovalsRepository.findByRequest_CustomerId(login.getId());
+            requests = requests.stream()
+                    .filter(request -> !requestApprovalsRepository.existsByRequestId(request.getId()))
+                    .collect(Collectors.toList());
         }
 
-        model.addAttribute("approvals", approvals);
+        for (Request request: requests) {
+            requestStatus.put(request.getId(), requestService.getStatus(request.getId()));
+        }
+
+        model.addAttribute("requests", requests);
+        model.addAttribute("requestStatus", requestStatus);
         return "my-reservations";
     }
 }
